refactor(JS): Moved the base page components into the application folder.
