+++ ui/src/main/js/application/App.js
@@ -3,19 +3,19 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
-import MainPage from 'MainPage';
-import LoginPage from 'LoginPage';
+import MainPage from 'application/MainPage';
+import LoginPage from 'application/LoginPage';
 import AboutInfoFooter from 'global/components/about/AboutInfoFooter';
 import { verifyLogin, verifySaml } from 'store/actions/session';
 import * as IconUtility from 'common/util/iconUtility';
-import LogoutPage from 'LogoutPage';
+import LogoutPage from 'application/LogoutPage';
 // These are needed for the react-bootstrap tables to show the ascending/descending icons
 import '@fortawesome/fontawesome-free/scss/fontawesome.scss';
 import '@fortawesome/fontawesome-free/js/all.js';
 import '@fortawesome/fontawesome-free/scss/v4-shims.scss';
 import '@fortawesome/fontawesome-free/js/v4-shims.js';
 
-import '../css/main.scss';
+import '../../css/main.scss';
 
 IconUtility.loadIconData();
